requirements.txt
langchain
python-dotenv==1.0.0
streamlit==1.22.0
unstructured==0.9.2
tiktoken==0.4.0
faiss-cpu==1.7.4
libmagic==1.0
python-magic
python-magic-bin
langchain_nvidia_ai_endpoints==0.1.2
sentence-transformers==2.2.2
transformers==4.33.3
langchain_community
langchain_groq
sentence_transformers